commits d26019 and 8ed15f from vanilla thunderstorm (fixes & cleanups)

+++ src/libs/Animals/Animals.cpp
@@ -31,7 +31,7 @@ bool ANIMALS::Init()
 uint64_t ANIMALS::ProcessMessage(MESSAGE &message)
 {
     const auto code = message.Long();
-    uint32_t outValue = 0;
+    uint64_t outValue = 0;
 
     switch (code)
     {
